pyproject.toml
[tool.poetry]
name = "kmat"
version = "0.1.0"
description = ""
authors = ["Rendy Arya Kemal <rendyarya22@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.4"
flask-migrate = "^2.5.3"
flask-login = "^0.5.0"
flask-admin = "^1.5.7"
authlib = "^0.15.2"
requests = "^2.25.0"
sentry-sdk = {version = "^0.19.3", extras = ["sentry"]}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
sqlalchemy_stubs = "^0.3"
flask-sqlalchemy-stubs = {git = "https://github.com/ssfdust/flask-sqlalchemy-stubs.git"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
